% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dina_vb.R
\name{dina}
\alias{dina}
\title{for deterministic input noisy AND gate (DINA) model.}
\usage{
dina(
  X,
  Q,
  max_it = 500,
  epsilon = 1e-99,
  delta_0 = rep(1, L),
  alpha_s = 1,
  beta_s = 1,
  alpha_g = 1,
  beta_g = 1
)
}
\arguments{
\item{Q}{J by K binary matrix, Q-matrix}

\item{max_it}{Maximum number of iterations}

\item{epsilon}{convergence tolerance for iterations}

\item{delta_0}{L by 1 vector, variational parameter of Dirichlet
distribution.}

\item{alpha_s}{Positive scalar, valvariational parameter that determines the
shape of  prior beta distribution for slip parameter \emph{s_j}.}

\item{beta_s}{Positive scalar, valvariational parameter that determines the
shape of  prior beta distribution for slip parameter \emph{s_j}.}

\item{alpha_g}{Positive scalar, valvariational parameter that determines the
shape of  prior beta distribution for guessing parameter \emph{g_j}.}

\item{beta_g}{Positive scalar, valvariational parameter that determines the
shape of  prior beta distribution for guessing parameter \emph{g_j}.}

\item{I}{by J binary matrix, item response data}
}
\value{
A list including:
\describe{
  \item{s_est}{the posterior mean of slip parameter.}
  \item{g_est}{the posterior mean of guessing parameter.}
  \item{s_sd}{the posterior standard diviation of slip parameter.}
  \item{g_sd}{the posterior standard diviation of guessing parameter.}
  \item{r_il}{the etimates of parameter for categorical distribution that is optimal variational posterior for \emph{\strong{z}_i}.}
  \item{alpha_s_ast}{the estimates of  for slip parameter}
  \item{beta_s_ast}{}
  \item{alpha_g_ast}{}
  \item{beta_g_ast}{}
  \item{pi_est}{}
  \item{delta_ast}{}
  \item{delta_sd}{}
  \item{l_lb}{}
  \item{att_pat_est}{}
  \item{A}{}
  \item{Q}{entered Q-matrix}
  \item{X}{entered data matrix}
  \item{eta_lj}{resulted idel responce}
  \item{m}{the number of performed iterations}
}
}
\description{
\code{dina()} returns variational Bayesian estimates for deterministic input
noisy AND gate (DINA) model.
}
\references{
Yamaguchi, K., & Okada, K. (2020). Variational Bayes inference
  for the DINA model. \emph{Journal of Educational and Behavioral
  Statistics}, 45(5), 569-597.
  \url{https://doi.org/10.3102/1076998620911934}.
}
