% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdina_vb.R
\name{mcdina}
\alias{mcdina}
\title{estimates attributes mastery pattern for saturated DCM.}
\usage{
mcdina(X, Q, max_it = 500, epsilon = 1e-05, seed = 123)
}
\arguments{
\item{X}{I by J matrix containing 0 or 1}

\item{Q}{J by K matrix containing 0 or 1}

\item{max_it}{Maximum number of iterations}

\item{epsilon}{convergence tolerance for iterations}

\item{seed}{seed value}
}
\description{
\code{mcdina()} returns variational Bayesian estimates for multiple-choice
DCM.
}
\references{
Yamaguchi, K. (2020). Variational Bayesian inference for the
  multiple-choice DINA model. \emph{Behaviormetrika}, 47(1), 159-187. \url{
  https://doi.org/10.1007/s41237-020-00104-w}
}
