% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satu_dcm_vb.R
\name{satu_dcm}
\alias{satu_dcm}
\title{estimates attributes mastery patterns for saturated DCM.}
\usage{
satu_dcm(X, Q, max_it = 500, epsilon = 1e-05, seed = 123)
}
\arguments{
\item{X}{I by J binary matrix, item response data}

\item{Q}{J by K binary matrix, Q-matrix}

\item{max_it}{Maximum number of iterations}

\item{epsilon}{convergence tolerance for iterations}

\item{seed}{seed value}
}
\description{
\code{satu_dcm()} returns variational Bayesian estimates for saturated DCM.
}
\references{
Yamaguchi, K., Okada, K. (2020). Variational Bayes Inference
  Algorithm for the Saturated Diagnostic Classification Model.
  \emph{Psychometrika}, 85(4), 973–995.
  \url{https://doi.org/10.1007/s11336-020-09739-w}.
}
